remove file lock.  don't make a backup file unless it exists.  very slight refactorings.  change case of project name / artifact to lowercase.

+++ src/main/java/com/martiansoftware/io/AtomicFileOutputStream.java
@@ -4,16 +4,14 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.RandomAccessFile;
-import java.nio.channels.FileLock;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 
 /**
  * This class is similar to the standard java.io.FileOutputStream, except that
- * the destination file is not touched until the stream is closed.  The destination
- * file is then atomically updated (replaced if necessary) with the freshly-written contents.
+ * all writes are directed to a temporary file, which then atomically replaces
+ * the destination file when the stream is closed.
  * 
  * The strategy is to:
  * <ol>
@@ -32,11 +30,10 @@ import java.nio.file.StandardCopyOption;
  */
 public class AtomicFileOutputStream extends OutputStream {
     
-    private final Path _dest;
-    private final Path _tmp;
-    private final Path _bak;
-    private final FileLock _fileLock;
-    private final FileOutputStream _out;
+    private final Path _dest; // user-specified destination file
+    private final Path _tmp;  // temporary file to which writes are redirected
+    private final Path _bak;  // backup of destination file made when stream is closed
+    private final FileOutputStream _tmpOut; // writes to tmp file
     
     /**
      * Creates a new AtomicFileOutputStream to write to or replace the specified File
@@ -54,13 +51,12 @@ public class AtomicFileOutputStream extends OutputStream {
      */
     public AtomicFileOutputStream(Path path) throws IOException {
         _dest = path;
-        _fileLock = new RandomAccessFile(_dest.toFile(), "rw").getChannel().tryLock();
         try {
             _tmp = Files.createTempFile(_dest.getParent(), 
-                                        _dest.getName(_dest.getNameCount() - 1).toString() + ".", 
+                                        _dest.getFileName() + ".",
                                         ".tmp");
             _bak = _tmp.resolveSibling(_tmp.getFileName().toString().replaceAll("\\.tmp$", ".bak"));            
-            _out = new FileOutputStream(_tmp.toFile());
+            _tmpOut = new FileOutputStream(_tmp.toFile());
         } catch (IOException e) {
             cleanup();
             throw e;
@@ -73,35 +69,17 @@ public class AtomicFileOutputStream extends OutputStream {
      */
     public Path getBackup() { return _bak; }
     
-    @Override public void close() throws IOException {
-        io(() -> _out.close());
-        io(() -> finish(true));
-    }    
-
+    /**
+     * Aborts the write process, leaving the destination file untouched
+     * @throws IOException 
+     */
     public void cancel() throws IOException {
-        io(() -> _out.close());
+        try { _tmpOut.close(); } catch (IOException ignored) {}
         io(() -> finish(false));
-    }
-
-    @Override public void flush() throws IOException {
-        io(() -> _out.flush());
-    }
-    
-    @Override public void write(byte[] b) throws IOException {
-        io(() -> _out.write(b));
-    }
-    
-    @Override public void write(byte[] b, int off, int len) throws IOException {
-        io(() -> _out.write(b, off, len));
-    }
-    
-    @Override public void write(int b) throws IOException {
-        io(() -> _out.write(b));
     }
     
     // wraps around an IO operation to clean up any temporary files if an exception occurs
     private void io(DelegatedIo d) throws IOException {
-        check();
         try { d.io(); }
         catch (IOException e) {
             cleanup();
@@ -111,11 +89,10 @@ public class AtomicFileOutputStream extends OutputStream {
 
     // perform the final operations to move the temporary file to its final destination
     private void finish(boolean success) throws IOException {
-        check();
         if (success) {
             try {
                 // copy original to .bak
-                Files.copy(_dest, _bak, StandardCopyOption.REPLACE_EXISTING);
+                if (Files.exists(_dest)) Files.copy(_dest, _bak, StandardCopyOption.REPLACE_EXISTING);
                 // atomically move tmp over original
                 Files.move(_tmp, _dest, StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
             } catch (IOException e) {
@@ -127,20 +104,36 @@ public class AtomicFileOutputStream extends OutputStream {
         Files.deleteIfExists(_bak); // only delete backup if everything works
     }
     
-    private void check() throws IOException {
-        if (!_fileLock.isValid()) throw new IOException("Stream is closed or invalid.");
-    }
-    
     private void cleanup() {
-        synchronized(_fileLock) {
-            if (_fileLock.isValid()) {
-                try { _fileLock.release(); } catch (IOException ignored) {}
-                try { Files.deleteIfExists(_tmp); } catch (IOException ignored) {}
-            }
-        }
+        try { Files.deleteIfExists(_tmp); } catch (IOException ignored) {}
     }
+
     
-    private interface DelegatedIo {
+    @FunctionalInterface private interface DelegatedIo {
         public void io() throws IOException;
     }
+
+
+    
+    @Override public void close() throws IOException {
+        io(() -> _tmpOut.close());
+        io(() -> finish(true));
+    }    
+
+    @Override public void flush() throws IOException {
+        io(() -> _tmpOut.flush());
+    }
+    
+    @Override public void write(byte[] b) throws IOException {
+        io(() -> _tmpOut.write(b));
+    }
+    
+    @Override public void write(byte[] b, int off, int len) throws IOException {
+        io(() -> _tmpOut.write(b, off, len));
+    }
+    
+    @Override public void write(int b) throws IOException {
+        io(() -> _tmpOut.write(b));
+    }    
+    
 }
